sampleapp/requirements.txt
amqp==2.5.2
billiard==3.6.1.0
boto3==1.10.46
botocore==1.13.46
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.0
django-allauth==0.41.0
django-appconf==1.0.3
django-crispy-forms==1.8.1
django-multiupload==0.5.2
django-rest-auth==0.9.5
django-storages==1.8
djangorestframework==3.11.0
docutils==0.15.2
idna==2.8
importlib-metadata==1.3.0
jmespath==0.9.4
kombu==4.6.7
more-itertools==8.0.2
oauthlib==3.1.0
Pillow==7.0.0
psycopg2==2.8.4
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
rcssmin==1.0.6
requests==2.22.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
